// api/src/constants/fazenda.ts
// eslint-disable-next-line import/prefer-default-export
export enum Fazenda {
  Id = '_id',
  Descricao = 'description',
  Nif = 'nif',
  Gerencia = 'gerencia',
  Contact = 'contact',
  Localization = 'localization',
  Estradanacional = 'estrada_nacional',
  DistanciaEstrada = 'distancia_estrada',
  Geo = 'geo',
  Ground = 'ground',
  ProducerId = 'producer_id',
  IsDeleted = 'is_deleted',
  Owner = 'onwer',
  Cultura = 'culturas',
  County = 'county',
  Machines = 'machines',
  Equipamentos = 'equipamentos',
  Infrastructure = 'insfrastructure',
  MeioEstacionario = 'meio_estacionario',
  Animals = 'animals',
  Extension = 'extension',
  HumanResource = 'human_resource',
}